requirements.txt
Django>=3.2,<4.0
PyMySQL>=1.0.2
python-dotenv>=0.19.0
gunicorn>=20.1.0
whitenoise>=5.3.0
